scraper/requirements.txt
requests==2.24.0
Scrapy==2.2.1
scrapy-splash==0.7.2
scrapyd==1.2.1
scrapyd-client==1.1.0
psycopg2-binary==2.8.5
